Trigger context menu create from background
