+++ test/main.test.js
@@ -8679,6 +8679,48 @@ describe("main", () => {
     });
   });
 
+  describe("handle contextmenu event", () => {
+    const func = mjs.handleContextmenuEvt;
+    beforeEach(() => {
+      if (typeof browser.menus.overrideContext !== "function") {
+        browser.menus.overrideContext = sinon.stub();
+      }
+    });
+
+    it("should call function", async () => {
+      const i = browser.menus.overrideContext.withArgs({
+        tabId: 1,
+        context: "tab",
+      }).callCount;
+      const elm = document.createElement("p");
+      const body = document.querySelector("body");
+      elm.dataset.tabId = "1";
+      body.appendChild(elm);
+      const res = await func({
+        target: elm,
+      });
+      assert.strictEqual(browser.menus.overrideContext.withArgs({
+        tabId: 1,
+        context: "tab",
+      }).callCount, i + 1, "called");
+      assert.isUndefined(res, "result");
+    });
+
+    it("should call function", async () => {
+      const i = browser.menus.overrideContext.withArgs({}).callCount;
+      const elm = document.createElement("p");
+      const body = document.querySelector("body");
+      elm.dataset.tabId = browser.tabs.TAB_ID_NONE;
+      body.appendChild(elm);
+      const res = await func({
+        target: elm,
+      });
+      assert.strictEqual(browser.menus.overrideContext.withArgs({}).callCount,
+                         i + 1, "called");
+      assert.isUndefined(res, "result");
+    });
+  });
+
   describe("handle runtime message", () => {
     const func = mjs.handleMsg;
